Settle issue #12
After letting this simmer for a bit, I've decided on removing the requirement for the *test* prefix for static dispatch test vi:s.

This means that any test utility method in public scope will be shown in the UI. However, if they do not contain assertions they will not cause failures. It will only look weird.
The fix to this is however straightforward - just restrict the access scope.